move common DataType2TfType() to utils.h
