+++ flutter/cpp/datasets/ade20k.cc
@@ -29,23 +29,6 @@ limitations under the License.
 
 namespace mlperf {
 namespace mobile {
-namespace {
-inline TfLiteType DataType2TfType(DataType::Type type) {
-  switch (type) {
-    case DataType::Float32:
-      return kTfLiteFloat32;
-    case DataType::Uint8:
-      return kTfLiteUInt8;
-    case DataType::Int8:
-      return kTfLiteInt8;
-    case DataType::Float16:
-      return kTfLiteFloat16;
-    default:
-      break;
-  }
-  return kTfLiteNoType;
-}
-}  // namespace
 
 ADE20K::ADE20K(Backend *backend, const std::string &image_dir,
                const std::string &ground_truth_dir, int num_classes,
